requirements.txt
streamlit>=1.28.0
openai>=1.3.0
anthropic>=0.7.0
boto3>=1.29.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0

